Move long running "addDel" test to TimingTests
